requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cron-validator==1.0.8
Django==4.2.4
django-celery-beat==2.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
flower==2.0.1
gunicorn==21.2.0
humanize==4.7.0
idna==3.4
kombu==5.3.1
packaging==23.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
